Move util functions to utils1.py
